# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.GetMeResponse:
    properties:
      avatar:
        type: string
      banner_image_url:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      profile_image_url:
        type: string
      public_id:
        type: string
      tier:
        type: string
    type: object
...
  dto.UpdateMeRequest:
    properties:
      email:
        type: string
      name:
        type: string
      settings:
        additionalProperties: true
        type: object
    type: object
...
  helpers.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@queuebuzz.com
    name: QueueBuzz Support
  description: Virtual queue management platform API.
  license:
    name: Proprietary
  termsOfService: https://queuebuzz.com/terms
  title: QueueBuzz API
  version: "1.0"
paths:
...
  /host/me:
    get:
      description: Gets the authenticated host profile.
      parameters:
      - description: Phone number
        in: query
        name: phone
        type: string
      - description: OTP code
        in: query
        name: otp
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Host profile
          schema:
            additionalProperties: true
            type: object
        "302":
          description: Found
          schema:
            allOf:
            - $ref: '#/definitions/helpers.SuccessResponse'
            - properties:
                Data:
                  $ref: '#/definitions/dto.GetMeResponse'
              type: object
        "400":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get authenticated host
      tags:
      - Host
    patch:
      description: Updates the authenticated host profile.
      parameters:
      - description: Host profile updates
        in: body
        name: host
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated Host profile
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update authenticated host
      tags:
      - Host
...
securityDefinitions:
  BearerAuth:
    description: 'RS256 JWT — format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
  UserToken:
    description: UUID user session token
    in: header
    name: X-User-Token
    type: apiKey
swagger: "2.0"